add: temp work, a lot of junk code, but some can be reused

+++ log_core/include/typedefs.hpp
@@ -1,5 +1,5 @@
 #ifndef TYPEDEFS_HPP
 #define TYPEDEFS_HPP
 typedef unsigned int RoomID;
-typedef unsigned long Timestamp;
+typedef unsigned int Timestamp;
 #endif
